update categories components
