+++ src/components/categories/list/table-body.jsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import Row from './table-row.jsx';
+
+const CategoriesTableBody = ({ categories }) => (
+  <tbody>
+    { categories.map((c) => <Row category={ c } key={ c.id } />) }
+  </tbody>
+);
+
+export default CategoriesTableBody;
